requirements.txt
websockets==11.0.3
uvicorn==0.21.1
starlette==0.26.1
requests==2.29.0
PyMySQL==1.0.3
pydantic==1.10.7
mysql-connector==2.2.9
Jinja2==3.1.2
loging==0.0.1
fastapi-sessions==0.3.2
fastapi-template==4.1.3
fastapi==0.95.1
python-decouple==3.8